MINOR: channel: add new function channel_congested()
This function returns non-zero if the channel is congested with data in
transit waiting for leaving, indicating to the caller that it should wait
for the reserve to be released before starting to process new data in
case it needs the ability to append data. This is meant to be used while
waiting for a clean response buffer before processing a request.
(cherry picked from commit 55e58f23348c2f9fa918e7370c0ddb98b307ee28)
(cherry picked from commit ac8771ccb2543777b19f2bde9140e3f41e93d11b)
[wt: channel_is_rewritable() is channel_reserved() in 1.5]